+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # [Releases](https://github.com/Tracktor/design-system/releases)
 
-## v4.31.6
+## v4.32.0
+
+### Breaking Changes
+- Remove `itemCount` from Kanban
 
 ### Fix 
-- Enhance infinite scroll logic in KanbanColumn to prevent redundant loading
+- Update KanbanColumn to adjust item count based on fetching state